requirements.txt
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==2.2.6
django-cors-headers==3.2.0
django-countries==6.1.2
django-filter==2.3.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
Markdown==3.2.2
moviepy==1.0.3
numpy==1.19.1
Pillow==6.2.1
proglog==0.1.9
PyJWT==1.7.1
pytz==2019.3
requests==2.24.0
sqlparse==0.3.0
tqdm==4.48.2
urllib3==1.25.10
zipp==3.1.0
